refactor: use components dir
