--- packages/atoms/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@rate/atoms",
-  "version": "0.0.1",
-  "files": [
-    "lib"
-  ],
-  "main": "lib/index.js",
-  "private": true,
-  "scripts": {
-    "build": "tsc -p .",
-    "clean": "rimraf lib",
-    "dev": "tsc -p . -w",
-    "prebuild": "yarn clean"
-  },
-  "typings": "lib/index.d.ts"
-}
