+++ packages/components/.gitignore
@@ -1,0 +1,5 @@
+/atoms/
+/molecules/
+/organisms/
+/pages/
+/templates/
