--- packages/molecules/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "@rate/molecules",
-  "version": "0.0.1",
-  "dependencies": {
-    "@rate/atoms": "0.0.1"
-  },
-  "files": [
-    "lib"
-  ],
-  "main": "lib/index.js",
-  "private": true,
-  "scripts": {
-    "build": "tsc -p .",
-    "clean": "rimraf lib",
-    "dev": "tsc -p . -w",
-    "prebuild": "yarn clean"
-  },
-  "typings": "lib/index.d.ts"
-}
