--- packages/molecules/src/index.ts
@@ -1,4 +1,0 @@
-export * from './Admire'
-export * from './AdmireContent'
-export * from './Navbar'
-export * from './Ranking'
