--- packages/pages/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@rate/pages",
-  "version": "0.0.1",
-  "files": [
-    "lib"
-  ],
-  "dependencies": {
-    "@rate/atoms": "0.0.1",
-    "@rate/templates": "0.0.1",
-    "@rate/molecules": "0.0.1"
-  },
-  "main": "lib/index.js",
-  "private": true,
-  "scripts": {
-    "build": "tsc -p .",
-    "clean": "rimraf lib",
-    "dev": "tsc -p . -w",
-    "prebuild": "yarn clean"
-  },
-  "typings": "lib/index.d.ts"
-}
