--- packages/templates/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@rate/templates",
-  "version": "0.0.1",
-  "dependencies": {
-    "@rate/atoms": "0.0.1",
-    "@rate/molecules": "0.0.1"
-  },
-  "files": [
-    "lib"
-  ],
-  "main": "lib/index.js",
-  "private": true,
-  "scripts": {
-    "build": "tsc -p .",
-    "clean": "rimraf lib",
-    "dev": "tsc -p . -w",
-    "prebuild": "yarn clean"
-  },
-  "typings": "lib/index.d.ts"
-}
