--- packages/templates/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./lib",
-    "noEmit": false,
-    "sourceMap": true,
-    "declaration": true
-  },
-  "exclude": ["**/*.test.ts", "**/*.test.tsx", "**/**/story.tsx", "lib", "es"]
-}
